ar51 - activerecord and other gems update
The Gemfile is completely up to date with this
commit. Not sure I recall ever seeing that before.
Seems like there is always one or two gems which
stubbornly refuse to update. I'm not actually sure
why it needs activerecord, now that I think of it.

Not much needs done with this project. Maybe get
it deployed in the wild at some point.

Medium coffee today. Will go with small coffee
tomorrow.

Beautiful day out there.